Other iterations/Base version/video-analytics/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Computer Vision & ML
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
insightface==0.7.3
onnxruntime==1.19.2
numpy==1.24.3
scikit-learn==1.3.2

# Image Processing
Pillow==10.1.0
imutils==0.5.4

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# Redis (for caching)
redis==5.0.1
hiredis==2.2.3

# WebSockets
websockets==12.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2

# API Documentation
pydantic[email]==2.5.0

# Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1

# Monitoring & Logging
loguru==0.7.2

# CORS
fastapi-cors==0.0.6
